Enable conversion of the key casing when converting objects.
The first ConvertOptions added is fromKeyCasing. This takes in the enum
key of the Casing enum.

Currently the following are the Casing enum:

* Camel
* Pascal
* Kabab
* Snake
* UpperSnake

To passing in the ConvertOptions, use the toConvert method. Example:

```
type User {
  userId: V.Uuid;
  userName: V.EmailAddress;
}

let isUser = V.isObject<User>({
  userId: V.isUuid,
  userName: V.isEmailAddress
})

isUser.toConvert({
  fromKeyCasing: 'Snake'
}).assert({
  user_id: '...', // an uuid
  user_name: '...' // an email address string.
});
```

Also - the typing of the convertOptions has been changed from ExplicitAny
to ConvertOptions.