Refactoring to fix constraint check failure
The test `S.isString.where((v) => v.length < 10).assert(null)` failed
with the weird error of constraint failed, when it should have failed
with null is not a string instead.

In order to solve this problem, I reworked the isa.js and convert.js to
refactor them into typescript instead of javascript to ensure that there
weren't any obvious type issues hidden in js. It took a while to get the
changes in, but after both are converted and correctly referenced, the
error went away. So yes there were some missing type issues in the JS.